Run Ansible from Virtual Environment
This changes updates the "apply-playbook.sh" script to execute Ansible
by installing a specific version into a "virtualenv" environment, and
then using the Ansible command from the virtual environment.

This should make the script more portable, as it isn't dependent on the
Ansible command installed on the system being configured. This also
allows the script to be run on a system where Ansible isn't installed,
or a version of Ansible is installed that is incompatible with the
playbook used by the script.